+++ example.ps1
@@ -26,3 +26,128 @@ New-MermaidGraph -Direction LR -NodeConnections @(
 ) -ClassDefinitions @(
     New-MermaidClassDefinition -Name "Starter" -FillColor "#6699ff" -StrokeColor "#999966"
 )
+
+
+$newMermaidClassDiagramSplat = @{
+    Class = @(
+        $(
+            $newMermaidClassSplat = @{
+                Name = 'Animal'
+                property = @(
+                    $(
+                        $newmermaidclasspropertySplat = @{
+                            Accessability = 'Public'
+                            Name = 'age'
+                            Datatype = 'int'
+                        }
+                        new-mermaidclassproperty @newmermaidclasspropertySplat
+                    ),
+                    $(
+                        $newmermaidclasspropertySplat = @{
+                            Accessability = 'Public'
+                            Name = 'Gender'
+                            Datatype = 'String'
+                        }
+                        new-mermaidclassproperty @newmermaidclasspropertySplat
+                    )
+                )
+                Method = @(
+                    $(
+                        $newMermaidClassMethodSplat = @{
+                            Encapsulation = 'Public'
+                            Name = 'isMammal'
+                        }
+                        New-MermaidClassMethod @newMermaidClassMethodSplat 
+                    ),
+                    $(
+                        $newMermaidClassMethodSplat = @{
+                            Encapsulation = 'Public'
+                            Name = 'mate'
+                        }
+                        New-MermaidClassMethod @newMermaidClassMethodSplat
+                    )
+                )
+            }
+            New-MermaidClass @newMermaidClassSplat
+        ),
+        $(
+            $newMermaidClassSplat = @{
+                Name = 'Fish'
+                property = @(
+                    $(
+                        $newmermaidclasspropertySplat = @{
+                            Accessability = 'Private'
+                            Name = 'sizeInFeet'
+                            Datatype = 'int'
+                        }
+                        new-mermaidclassproperty @newmermaidclasspropertySplat
+                    )
+                )
+                Method = @(
+                    $(
+                        $newMermaidClassMethodSplat = @{
+                            Encapsulation = 'Private'
+                            Name = 'canEat'
+                        }
+                        New-MermaidClassMethod @newMermaidClassMethodSplat 
+                    )
+                )
+            }
+            New-MermaidClass @newMermaidClassSplat
+        ),
+        $(
+            $newMermaidClassSplat = @{
+                Name = 'Duck'
+                property = @(
+                    $(
+                        $newmermaidclasspropertySplat = @{
+                            Accessability = 'Public'
+                            Name = 'beackColor'
+                            Datatype = 'string'
+                        }
+                        new-mermaidclassproperty @newmermaidclasspropertySplat
+                    )
+                )
+                Method = @(
+                    $(
+                        $newMermaidClassMethodSplat = @{
+                            Encapsulation = 'Public'
+                            Name = 'swim'
+                        }
+                        New-MermaidClassMethod @newMermaidClassMethodSplat 
+                    ),
+                    $(
+                        $newMermaidClassMethodSplat = @{
+                            Encapsulation = 'Public'
+                            Name = 'quack'
+                        }
+                        New-MermaidClassMethod @newMermaidClassMethodSplat
+                    )
+                )
+            }
+            New-MermaidClass @newMermaidClassSplat
+        )
+    )
+    RelationShip = @(
+        $(
+            $newMermaidClassRelationShipSplat = @{
+                RelationShipType = 'Inheritance'
+                FirstClass = 'Animal'
+                SecondClass = 'Duck'
+            }
+
+            New-MermaidClassRelationShip @newMermaidClassRelationShipSplat
+        ),
+        $(
+            $newMermaidClassRelationShipSplat = @{
+                RelationShipType = 'Inheritance'
+                FirstClass = 'Animal'
+                SecondClass = 'Fish'
+            }
+
+            New-MermaidClassRelationShip @newMermaidClassRelationShipSplat
+        )
+    )
+}
+
+New-MermaidClassDiagram @newMermaidClassDiagramSplat